extract 'withSVGTrail' HOC
