+++ src/components/LayoutComponents/index.tsx
@@ -2,4 +2,5 @@ export { default as MobileDrawer } from "./MobileDrawer"
 export { default as NavBar } from "./NavBar"
 export { default as Footer } from "./Footer"
 export { default as Header } from "./Header"
+export { default as withSVGTrail } from "./HOCs/withSVGTrail"
 export { styles } from "./Styles"
